Add tx building for MegaSwap
Refactor payload encoding
